POLIO-839: replace custom code with mixin
